+++ server/src/resolvers/Message.ts
@@ -15,11 +15,21 @@ export class MessageResolver {
   ): Promise<Message> {
     try {
       // create the new message
-      return await Message.create({
+      const newMessage = await Message.create({
         content: message,
         userId,
         conversationUuid: convUuid,
       }).save();
+
+      return await getConnection()
+        .getRepository(Message)
+        .createQueryBuilder("message")
+        .leftJoinAndSelect("message.user", "user")
+        .leftJoinAndSelect("user.service", "service")
+        .leftJoinAndSelect("user.poste", "poste")
+        .where("message.conversationUuid = :convUuid", { convUuid })
+        .andWhere("message.uuid = :uuid", { uuid: newMessage.uuid })
+        .getOneOrFail();
     } catch (err) {
       return err;
     }